+++ src/js/lang/default.js
@@ -1,6 +1,8 @@
 export default {
     moveLeft: 'Move left',
-    moveRight: 'Move right',
     moveAllLeft: 'Move all left',
+    moveRight: 'Move right',
     moveAllRight: 'Move all right',
+    moveDown: 'Move item down',
+    moveUp: 'Move item up',
 };