+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features
 
+* [#24]: Add `showOrderButtons` property to allow users to move the selected options up or down in the list
 * [#48]: Add `allowDuplicates` property to allow more than one copy of an available value to be selected
 * [#50]: Add `id` property for control over the HTML ID assigned to the component instance
 * [#54]: Add `lang` property for localization